requirements.txt
RPi.gpio
flask
flask-api
pyserial
adafruit-circuitpython-servokit